Refactored cursor management files
